host/bert_ner/requirements.txt
python-dotenv
gradio
